+++ src/js/Content/Features/Community/CCommunityBase.js
@@ -1,5 +1,6 @@
 import {Context, ContextType} from "../../modulesContent";
 import FHideTrademarks from "../Common/FHideTrademarks";
+import FBackToTop from "../Common/FBackToTop";
 import FConfirmDeleteComment from "./FConfirmDeleteComment";
 import FHideSpamComments from "./FHideSpamComments";
 import FFavoriteEmoticons from "./FFavoriteEmoticons";
@@ -10,6 +11,7 @@ export class CCommunityBase extends Context {
 
         features.push(
             FHideTrademarks,
+            FBackToTop,
             FConfirmDeleteComment,
             FHideSpamComments,
             FFavoriteEmoticons,
